% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_test.R
\name{heart_statistic}
\alias{heart_statistic}
\title{HEART test statistic}
\usage{
heart_statistic(p_tar, p_de, par, log = FALSE)
}
\arguments{
\item{p_tar, p_de}{Target-only and decorrelated p-values.}

\item{par}{List with \code{pi00}, \code{pi01}, \code{alpha1}, \code{alpha2}, \code{C1}, \code{C2}
(as returned in \verb{$par} by \code{\link[=heart_test]{heart_test()}}).}

\item{log}{Return \eqn{\log T_j} (recommended for very small p-values).}
}
\value{
Numeric vector of statistics.
}
\description{
Computes the simplified HEART statistic (with \eqn{\pi_{10}} set to zero)
\deqn{T_j = \hat\pi_{00}\hat C_1^{-1}\hat C_2^{-1}\hat\alpha_1^{-1}\hat\alpha_2^{-1}
  p_{tar,j}^{1-\hat\alpha_1} p_{de,j}^{1-\hat\alpha_2} +
  \hat\pi_{01}\hat C_1^{-1}\hat\alpha_1^{-1} p_{tar,j}^{1-\hat\alpha_1}.}
Small values are evidence against \eqn{H_{0,j}: \beta^{tar}_j = 0}.
}
