% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{heart}
\alias{heart}
\title{Title HEART: A high-efficiency and robust test for large scale
single-cell RNA-sequencing differential expression analysis}
\usage{
heart(us.data, ln = F, mt = "holm", p.thr = 0.01)
}
\arguments{
\item{us.data}{input data. It is a list with 2 elements,
The first element names 'counts', us.data$counts should be  a data.frame or a matrix,
 where rows are genes and columns are cells.
The second element names 'cell.data'. us.data$cell.data must be a data.frame,
and it contains at least two columns,
one of the column of us.data$cell.data should be  named 'cell.id', which corresponds to the column name in us.data$counts,
the other one of the column of us.data$cell.data should be named 'group',
which is a pre-defined group/class of each cell

Additionally:

1. The us.data$counts must have the same number of columns as the us.data$cell.data has rows.
2. The us.data$cell.dta$cell.id should match the column names of the us.data$counts.}

\item{ln}{logical value; default is FALSE.
If TRUE, then the data will compute the log2 transformation,}

\item{mt}{Character string, FDR methods, default is 'holm'.}

\item{p.thr}{Numeric value.default is 0.01.}
}
\value{
a list with 2 elements.
The first element named 'end.out', which contains identified DE gene and corresponding p-values and adjusted p-values
The second element of the output list named 'all.statistics', which contains statistics in the calculation process.
}
\description{
Title HEART: A high-efficiency and robust test for large scale
single-cell RNA-sequencing differential expression analysis
}
\examples{



}
